Initial SetUp
I've commenced the project by establishing the foundational elements. Utilising MySQL Workbench, the database setup ensures a seamless connection with the application. The server-side code, along with its dependencies, is meticulously crafted to facilitate RESTful API interactions for data posting and retrieval. The database now contains information on the current Premier League teams, laying the groundwork for users to cast their votes on the anticipated winner of the 2024 Premier League. This initial setup positions the project for a professional and interactive user experience.