Completed functions
Current implementation is functional, as all key requirements are met, such as submitting votes, to retrieving the results. However the focus on local storage is extremely heavy on my client side. To resolve this, on the result side, obtaining the results on the server side, passing it through on the client side and displaying the results this way will be done