Updated Welcome and Voting page
I had reread the task and changed my current functions to then utilise the server side. Rather than directly the local storage direct front the client side, it is now incorporated that the serverSide and the client side work alongside each other to send the results to local storage. The design for the welcome page and voting page are now complete. The results page is left to complete along with its design.